Revert the revert of "Use common test build during CI (#7196)" (#7404) (#7870)
This reverts commit f9bdd102088dad6c02bbfb0c8b3af2c16b708ab1. The
original change was a possible optimization of CI, though it ended up
not having a huge impact. It was thought that it may have broken source
maps, because the test build overwrote the `dist` directory that the
source maps were written to. However this turned out not to be the
case, as the changes to `dist` are never persisted to the workspace.

This is being re-introduced because the test build is needed for an
additional job (the page load benchmark), and sharing the same build
for all three jobs would surely be faster than building it separately
three times.