feat: CircleCI is now storing builds-test and builds-test-flask (#20749)
## Explanation

The intent of this PR is very similar to #20783. We want to save a
little time for QA and developers, when they have to write, modify, or
debug local E2E tests. The `builds-test` and the `builds-test-flask` can
be downloaded from the PR instead of locally generated. It also ensures
that your local E2E is running the same code as the CircleCI E2E.

The first version of this PR had to do some acrobatics with `save_cache`
because `prep-build-test-flask` was not connected to
`jobs-publish-prerelease`. But now they are connected, so this PR is now
more straightforward.

## Screenshots/Screencaps

### Before


![image](https://github.com/MetaMask/metamask-extension/assets/539738/7f6ccba5-8abe-422c-ad02-9b119df2dd66)

### After


![image](https://github.com/MetaMask/metamask-extension/assets/539738/cf52f65f-be60-41ef-a667-a218968fbc07)

## Manual Testing Steps

- For all 4 builds:
  - builds (test): chrome, firefox
  - builds (test-flask): chrome, firefox
1. Download them from the latest metamaskbot links on this PR
2. Open their manifest.json files to confirm that they have
`"description": "testing build from git id: xxxx",`
and `"name": "MetaMask Main lavamoat snow",` or `"name": "MetaMask Flask
lavamoat snow",`
3. Try them with `yarn test:e2e:single`, make sure they run E2E tests
properly

The QA Engineer Mike Berardi was already testing this as I was
developing it.

## Pre-merge author checklist

- [x] I've clearly explained:
  - [x] What problem this PR is solving
  - [x] How this problem was solved
  - [x] How reviewers can test my changes
- [x] Sufficient automated test coverage has been added

## Pre-merge reviewer checklist

- [X] Manual testing (e.g. pull and build branch, run in browser, test
code being changed)
- [ ] PR is linked to the appropriate GitHub issue
- [ ] **IF** this PR fixes a bug in the release milestone, add this PR
to the release milestone