Ignored ripple-lib deprecation advisory, as we do not depend on it in MetaMask (#21182)
## **Description**
This PR addresses a `yarn audit` failure that we currently see on
develop. The failure is due to the deprecation of the `ripple-lib`
library used within `@trezor/blockchain-link@npm:2.1.8`. Although
imported in one of our dependencies, MetaMask does not depend on that
library, and it's code is not used by our codebase. We can safely ignore
this advisory.

## **Manual testing steps**

Not needed. Builds should pass without a `yarn audit` warning