Fix/ap 63/add message to loading for hw and snap accounts (#21229)
## **Description**

This PR adds a loading message to the loading indicator for Hardware
Wallet and Snap Accounts. Previously, when a transaction was made and
not confirmed on the device, MetaMask would stay in a loading state
indefinitely, which could be confusing for users. With this change, a
loading message is displayed to indicate that the transaction is waiting
for confirmation on the device. This improves the user experience and
makes it clear that the transaction is still in progress.

## **Manual testing steps**

1. Using a Snap (using async flow) or HW account 
2. Click send
3. Send any amount of eth and confirm transaction
4. Don't approve the transaction on the hardware device or snap account
5. See the new loading message

## **Screenshots/Recordings**

### **Before**


![snap-loading-screen2](https://github.com/MetaMask/metamask-extension/assets/96463427/9fadbca7-682a-4ace-81b3-9b62289f453e)

### **After**


![snap-loading-screen](https://github.com/MetaMask/metamask-extension/assets/96463427/5b035bdd-a345-4814-9a34-739298e37378)

## **Related issues**

Resolves https://github.com/MetaMask/accounts-planning/issues/63

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've clearly explained:
  - [x] What problem this PR is solving.
  - [x] How this problem was solved.
  - [x] How reviewers can test my changes.
- [x] I’ve indicated what issue this PR is linked to: Fixes #???
- [ ] I’ve included tests if applicable.
- [ ] I’ve documented any added code.
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
- [x] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [ ] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [x] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [x] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.