Bump snaps packages (#21255)
## **Description**
Bump snaps packages to latest. 

This release adds some under the hood support for a new keyring export.
It also adds support for showing SVGs in snaps custom UI.

As part of bumping the packages we had to update TypeScript as well. We
had a long-standing patch to TypeScript that could be easily replaced
with a small patch to LavaMoat to enable `overrideTaming`. This feature
is also set to be enabled in the next version of LavaMoat.