fix: After triggering a Transaction, a Signature request metrics event is also triggered (#21753)
## **Description**

Problem: after triggering any transaction (not a signature) we can see
how now a Signature request is also triggered. Additionally, the
information about RPC requests is held in the Signature request (which
should not be there) instead of in the transaction request.

## **Expected behavior**
No Signature event should be triggered if we perform a transaction
RPC information should be held inside the transaction request, instead
of the signature, for transactions

## **Fix**
We are sending signature events on confirm page, we shouldn't. This call
is now removed. Then on Transaction Events, we should add the blockaid
RPC calls. There's already a code for this in metrics helper. We just
need to call this in metrics to add the blockaid params.

## **Related issues**

Fixes: #21739 

## **Manual testing steps**

1. Enable a metrics project / or look the background console
2. Select Mainnet
3. Go to the test dapp
4. Trigger a transaction
5. See how no info about RPC is held in the transaction event
6. See how a Signature request is also triggered, alongside the
transaction
7. Checkout this branch
8. Repeat step 1-4
9. See how RPC info is sent in Transaction Added event
10. See how no signature request event is triggered.

## **Screenshots/Recordings**

### **Before**

https://github.com/MetaMask/metamask-extension/assets/54408225/96293fc6-507c-41d1-9fbc-e38a011ff845

### **After**


https://github.com/MetaMask/metamask-extension/assets/44811/44bbd5f6-247a-4369-8b2c-9f3c8c180075


## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've clearly explained what problem this PR is solving and how it
is solved.
- [x] I've linked related issues
- [x] I've included manual testing steps
- [x] I've included screenshots/recordings if applicable
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [x] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [x] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Signed-off-by: Akintayo A. Olusegun <akintayo.segun@gmail.com>