feat(mv3): Ledger MV3 support (#22275)
## **Description**

Adds the ledger-iframe to the offscreen document and establishes MV3
support for ledger. This is done by the following method:

1. ledger-iframe.ts interfaces with the ledger keyring iframe, working
as a middle layer between the two. Messages received by the offscreen
document that target the ledger iframe will be forwarded to the ledger
keyring iframe (the github pages generated content from
@MetaMask/eth-ledger-bridge-keyring. A callback manager uses an
incremental id to assign an id to each message which is returned in the
response such that a callback can be declared only for a specific id.
2. ledger-offscreen-bridge.ts is a bridge for the LedgerKeyring that
interfaces with the offscreen document such that when the keyring calls
out for specific methods, it calls into the offscreen bridge and then
the communication is sent to the offscreen document and forwarded to the
keyring iframe.

In addition some more cleanup of the #18794 deprecation of u2f and
ledger live. We no longer support setting any other method than the
browsers lowest level method for connecting. For chrome webhid is the
gold standard, for firefox which doesn't support webhid we use u2f.
There were still a number of places in the code that reference the
transportType from preferences and migrations were done to change the
value for many users when we only need to know what the support for
those two protocols are.

So firstly it changes the setLedgerTransportPreference to not take an
input, but rather just uses the browsers support of navigator.hid to
determine whether to use webhid or u2f. When #18794 landed service
workers did not have access to webhid, now they do, so this is a viable
approach. I also added deprecation messages where appropriate. In a
future PR we should go through and remove any usages of the preference.
Ideally we'd also set the transport type for the ledger device closer to
the bridge, however i'm not familiar with the unlock mechanism of the
keyring and trying to do this was causing issues with the mv3
connectivity.

## **Related issues**

Fixes #17724 - No longer relevant because background/service workers
have access to window.navigator.webhid
Fixes #21624 - Connects ledger to offscreen doc

## **Screenshots/Recordings**



## **Manual testing steps**

1. run `yarn start:mv3`
2. reload the extension in your browser.
3. try to add a ledger device. (Remember you must have your device
unlocked and the ethereum app open on your ledger
4. Try to use the test dapp to send tx, sign messages, etc.

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've clearly explained what problem this PR is solving and how it
is solved.
- [x] I've linked related issues
- [x] I've included manual testing steps
- [x] I've included screenshots/recordings if applicable
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [x] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [x] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.