Fix issue connecting to dapps that dont use eth_accounts (#22405)
## **Description**

Fixes an issue with connecting to dapps that have permissions (e.g. for
snaps) but don't leverage `eth_accounts`.

A flawed assumption was made in which we assumed that dapps that have
permissions always have `eth_accounts`, that is not the case. The
applied fix bails out of the function in question if `eth_accounts` is
not present as it is expected to be.

## **Manual testing steps**

1. Disconnect from `metamask.github.io`
2. Open https://metamask.github.io/snaps/test-snaps/latest/
3. Install a snap
4. Refresh the page and notice that you can still use the dapp