E2E test: Improve fill method in driver (#22652)
## **Description**

When investigating flaky tests, I found that some tests are flaky
because the input field was not cleared before entering a new value. The
issue is that Selenium sometimes doesn't clear input fields effectively.
Now, we use sendkeys() to clear input fields, i've added another way to
use selenium actions to simulate we select all text with keyboard and
replace it with the new value. With both ways, we can clear input fields
more effectively, which I believe will stabilize multiple flaky tests.

Examples:

Failing test because input field was not cleared as expected:
![Screenshot 2024-01-24 at 11 46
01](https://github.com/MetaMask/metamask-extension/assets/105063779/5f067bd0-c426-42fd-b378-62b6dbba6b8a)


Tests failing because although the input field was cleared, the
application automatically added a '0'. So when we entered the input
value '4000', it was actually entered as '04000'. With my approach, when
we find that the input field is not empty, it will select the '0' and
replace it with '4000'. This way, we finally get the correct value of
'4000'.
<img width="363" alt="Screenshot 2024-01-24 at 14 44 04"
src="https://github.com/MetaMask/metamask-extension/assets/105063779/98d6d811-46df-4d15-be50-7f18fe13fdf7">


I also fixed another flaky test in this PR.

## **Related issues**

Fixes: #22573 

## **Manual testing steps**

Run the e2e test.