Petnames: Addition of shouldSyncPetname Method for Petnames Synchronization Control (#22600)
## **Description**

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->
This PR introduces the `shouldSyncPetname` method to
`AbstractPetnamesBridge` to provide a refined control mechanism over
which Petname entries participate in synchronization with another
source.

By default, every Petname entry is automatically considered for
synchronization, ensuring that the Petname state is a comprehensive
mirror of the source entries and vice versa post-synchronization. By
overriding this method, developers can now specify which Petname entries
should be included or excluded from the synchronization process.

This has implications, depending on the sync direction:

- `Source->Petnames`: Entries in the Petnames state, which are marked as
non-participants (i.e., shouldSyncPetname returns false), will not be
deleted during the synchronization process. This ensures that specific
entries are preserved in the Petnames state, regardless of their absence
or deletion in the source.
- `Petnames->Source`: In this synchronization direction, entries in the
Petnames state marked as non-participants will not be added to the
source. This allows for maintaining entries exclusively in the Petnames
state without affecting the source dataset.

## **Related issues**

* Fixes: https://github.com/MetaMask/metamask-extension/issues/22556

## **Manual testing steps**

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've clearly explained what problem this PR is solving and how it
is solved.
- [ ] I've linked related issues
- [ ] I've included manual testing steps
- [ ] I've included screenshots/recordings if applicable
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [ ] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [ ] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: MetaMask Bot <metamaskbot@users.noreply.github.com>
Co-authored-by: chloeYue <105063779+chloeYue@users.noreply.github.com>
Co-authored-by: Harsh Shukla <125105825+PrgrmrHarshShukla@users.noreply.github.com>
Co-authored-by: Olusegun Akintayo <akintayo.segun@gmail.com>
Co-authored-by: Ariella Vu <20778143+digiwand@users.noreply.github.com>
Co-authored-by: David Walsh <davidwalsh83@gmail.com>
Co-authored-by: Monte Lai <monte.lai@consensys.net>
Co-authored-by: Sébastien Van Eyck <sebastien.vaneyck@consensys.net>
Co-authored-by: Mark Stacey <markjstacey@gmail.com>
Co-authored-by: legobeat <109787230+legobeat@users.noreply.github.com>
Co-authored-by: Mako Shan <imakoshan@gmail.com>
Co-authored-by: Sébastien Van Eyck <sebastien.vaneyck@gmail.com>
Co-authored-by: salimtb <salim.toubal@outlook.com>
Co-authored-by: gantunesr <gustavoantunesrod@gmail.com>
Co-authored-by: Gustavo Antunes <17601467+gantunesr@users.noreply.github.com>
Co-authored-by: Howard Braham <howrad@gmail.com>
Co-authored-by: Tomáš Martykán <t.martykan@gmail.com>
Co-authored-by: Brad Decker <bhdecker84@gmail.com>
Co-authored-by: Daniel <80175477+dan437@users.noreply.github.com>