Simplify permissions page logic (#22850)
## **Description**
Removes a bunch of outdated logic from the permissions page that is not
in use anymore.

We previously stored the requested permissions alongside a boolean of
whether they should be granted, but the option to toggle that boolean
was removed years ago. This PR replaces that logic with deriving the
requested permissions from the props without storing them in state. This
also simplifies later iterations where the permission values (not just
the keys) are needed to render the permissions.


## **Manual testing steps**

1. Go to the test-dapp
2. Connect to the test-dapp
3. See that it still works