Cherry-pick 8dc1139 (#22978) to v11.9.5
The issue described here
https://github.com/MetaMask/metamask-mobile/issues/8589 is also
happening in the extension (in prod) in v11.9.x

The root of the problem lies in a geth update and how a
`eth_createAccessList` request is handled when passed certain variables.

The incompatibility of ppom and the latest changes in geth is handled in
v1.4.1 of `ppom_release`.

Fixes:

1. Go to https://metamask.github.io/test-dapp/ and connect your wallet
2. Click "Malicious ERC-20 Transfer"
3. A confirmation popup should appear with a warning like the one shown
below

![Screenshot from 2024-02-15
18-14-17](https://github.com/MetaMask/metamask-extension/assets/7499938/cca14ca0-eaa2-4e93-9112-d170bef2658e)

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

https://github.com/MetaMask/metamask-extension/assets/54408225/883b58c8-feb3-4e20-9013-2b54d14de68c

https://github.com/MetaMask/metamask-extension/assets/54408225/cfc5e10d-ff25-42bb-ace2-052452611e14

- [ ] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've clearly explained what problem this PR is solving and how it
is solved.
- [ ] I've linked related issues
- [ ] I've included manual testing steps
- [ ] I've included screenshots/recordings if applicable
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [ ] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [ ] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: Dan J Miller <danjm.com@gmail.com>