fix: duplicate React keys in dynamic UI causing artifacts (#23139)
## **Description**

This PR fixes two problems that cause the same bug. We are currently not
effectively generating unique keys for components used in Snaps custom
UI. This becomes a problem especially when using dynamic UI and
re-rendering the screen where React is unable to clean up the previous
screen and may fail to remove old components.

The fix is two-fold: First, we need to always have a unique suffix
("root key") for our keys, so that changing the UI content always
re-renders the entire screen and cleans out old components. Secondly, we
need to restore a trick that was used before recent changes to
`elementKeyIndex`. When using an object container for `elementKeyIndex`
we can reference the same number value across recursion. In a recent PR
this trick was removed in favor of simply passing a number, but passing
the number passes the literal value and thus results in duplicates keys
in potentially complex component trees.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/23139?quickstart=1)