test: update tests with types, include tests in linting (#23136)
https://github.com/MetaMask/metamask-extension/pull/22530 broke eslint
for `*.test.ts` and `*.test.tsx` files. This PR restores eslint
functionality in those files.

In order for eslint to work with typescript files the typescript files
need to have a corresponding `tsconfig.json`. Instead of having a
separate config for our test files I've added the test files to our
existing `tsconfig.json`. Many of the tests didn't pass the type
checker, so I had to update types in the tests to get everything
working.

This PR does not restore eslint for `*.stories.ts(x)` files. I think
that should be in another PR, as there are 40 stories files that don't
pass the type checker.