chore: update `sass` to v1.71.0 (#22531)
## Updates Sass

Updates `sass` to the latest version, while also switching to
`sass-embedded` (maintained by the same team, is faster than sass, and
we don't need to patch it anymore!).

## Moves fonts

This PR also moves the `fonts` directory to be relative to the CSS file
that imports it. This helps with two things:

1. its just nice to use _real_ relative paths to files that are using
them while developing (not just in the final build)
2. the webpack build will work differently: it _requires_ all the
imports to use relative paths (there is a hack I can put in place to not
require relative paths, but meh. I don't want to do that :-) ).

## Use `@use` for the design-system instead of `@import`

The biggest change in this PR is moving towards the use of `@use` and
away from `@import`, which is slated to be completely removed from Sass
in some distance future release.

### Bonus savings!

This PR also fixes a the duplication of CSS, as the design-system
previously included actual CSS (colors, fonts, and font-awesome styles)
and reduces the output size from ~1700KB to ~1300KB. The issue with that
some devs were already `@use`ing the `design-system`, which was
re-including its all of its rules (mentioned above).

### ⚠️ Changes How We Write SCSS

So, from now on, if someone wants to use something from the
design-system they'll need to put `@use design-system";` at the top of
the SCSS file.