Feat(MV3): Support Snaps in MV3 (#22732)
## **Description**
The work done in the proceeding offscreen PRs was done explicitly to
hopefully make adopting the snaps executor written for offscreen support
without much additional effort. This PR is not complete because it has a
type error but sharing the code changes now.

This PR pulls forward the ProxySnapsExecutor from the
snaps-execution-environments repository and initializes it with a
runtime message stream. This allows the snaps executor to communicate
through the offscreen document and enable support for snaps in MV3

## **Related issues**

[MetaMask-planning#1998](https://github.com/MetaMask/MetaMask-planning/issues/1998)

## **Manual testing steps**

1. Pull branch
2. run `yarn`
3. run `yarn start:mv3`
4. Go to https://metamask.github.io/snaps/test-snaps/latest/
5. Install some of the snaps by clicking "reconnect to X"
6. Interact with the buttons after installing snap and see they work.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've clearly explained what problem this PR is solving and how it
is solved.
- [x] I've linked related issues
- [x] I've included manual testing steps
- [x] I've included screenshots/recordings if applicable
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [x] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [ ] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: MetaMask Bot <metamaskbot@users.noreply.github.com>