fix: Can't call setApprovalForAll on certain NFTs (#22822)
## **Description**

Fixes an issue where some NFTs can't be listed because MetaMask crashes
with:
`Invalid data; should be 'approve' method, but instead is
'setApprovalForAll'`

This happens when MetaMask categorizes the NFT as an ERC20 token. A dapp
prompts `setApprovalForAll` (from ERC721/ERC1155), but MetaMask
mistakenly enters a flow designed for ERC20 `approve`.

With specs like ERC404 that are intentionally fungible and non-fungible,
MetaMask should allow calling `setApprovalForAll` on ERC20-like tokens.

The fix is to enter the `setApprovalForAll` flow when dapps prompt that
function signature, even if metamask categorizes the token as an ERC20.

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-extension/issues/18507

## **Manual testing steps**

It may be hard to obtain a token of this type, but you could try to list
an ERC404 token on OpenSea.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**


https://github.com/MetaMask/metamask-extension/assets/3500406/d8342357-0eaf-41fd-bac2-71b58d3a34df


<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've clearly explained what problem this PR is solving and how it
is solved.
- [ ] I've linked related issues
- [ ] I've included manual testing steps
- [ ] I've included screenshots/recordings if applicable
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [ ] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [ ] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.