chore(ts): Use FirstTimeFlowType enum (#23501)
## **Description**
As part of #23056 we discovered that it will be necessary to introduce a
new "type" of firstTimeFlow which is 'restore'. In order to reason about
those changes I wanted to shift to using an enum for FirstTimeFlowType
which can actually be null, 'import' or 'create' at present.
@NiranjanaBinoy converted the onboarding controller to typescript and
introduced the enum in
https://github.com/MetaMask/metamask-extension/pull/23461 and this PR
propogates its usages.

## **Related issues**

## **Manual testing steps**
1. There should be no testing required as this is not a functional
change. However, validating that onboarding still works as intended is
viable, and hopefully any bugs will be severe enough to be caught In CI.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've clearly explained what problem this PR is solving and how it
is solved.
- [x] I've linked related issues
- [x] I've included manual testing steps
- [x] I've included screenshots/recordings if applicable
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [x] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [x] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.