fix: updated connections page for connected data (#24022)
In the previous logic, we were using activeTabOrigin to get the subject
MetaData to get the connected Dapp MetaData but it seems like there is
another bug because of which metaData is not being fetched correctly. As
a result, we are not able to get the connected Dapp Logo and connected
status properly. In this PR, we are relying on
getPermittedAccountsForCurrentTab to get the connected status and if the
connected dapp metaData is not defined, we will rely on global icon
rather than connected dapp icon

## **Related issues**

Fixes: #23792 

## **Manual testing steps**

1. Run extension with Multichain Flag and connect it to dapp
2. If the dapp shows connected status, then the connection page should
return the connected accounts
3. Test it with multiple Dapps

## **Screenshots/Recordings**


### **Before**
Refer to the issue

### **After**



https://github.com/MetaMask/metamask-extension/assets/39872794/4536a206-a4de-45e4-88ca-cdbaf92478bb




## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.