fix: Updated selector for the test networks (#23986)
This PR is to ensure that the permissions page renders list of all the
permissions with the current networks. For a few networks, we don't have
the rpcPrefs that was causing the siteKey nd the rpcPrefs to be
undefined.
This PR also makes sure that the defined background color is there for
the test networks that doesn't have a logo


## **Related issues**

Fixes: #23812 #23793 


## **Manual testing steps**

1. Connect an account on Uniswap
2. Open the popup
3. Open the three-dot menu and choose "All Permissions"
4. It should render the permission with the accounts

## **Screenshots/Recordings**


### **Before**


https://github.com/MetaMask/metamask-extension/assets/39872794/4fc1e04c-fde3-4905-a8b8-b16c75d9f7d1



### **After**


https://github.com/MetaMask/metamask-extension/assets/39872794/c56a60cb-24e6-4857-ac29-2539bb0f6645



## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.