fix: fix migration 105 to recover from undefined selected address (#23974)
## **Description**

This pull request fixes the selectedAddress migration issue where the
selectedAddress is not a string when migrating to internal accounts. In
this broken state, the first identity would then be used for the
`selectedInternalAccount` and `selectedAddress`

## **Related issues**

Fixes: [383](https://github.com/MetaMask/accounts-planning/issues/383)

## **Manual testing steps**

1 . Comment out all the migrations from 105 and onwards.
2. Reinstall the extension and create 2 accounts, and select the second
account. Dump the metamask state and see that the second account is the
selectedAddress
3. In migration 105, force the selectedAddress to be undefined.
4. Uncomment the migrations
5. Restart the extension, see that the selectedAddress is now the first
account

## **Screenshots/Recordings**

### **Before**

### **After**

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.