feat: push notifications new controller (#23137)
## **Description**

This PR introduces a new SW (Service Worker) and a new controller for
managing push notifications.

This integration, due to the differences between MV2 and MV3 regarding
the use of background scripts and the SW, is only related to the
manifest V2. For the manifest V3, part of this logic will be modified,
using the extension's own service worker as a starting point. With
manifest V2, on the contrary, it is necessary to initialize a dedicated
service worker so that FCM (Firebase Cloud Messaging) can handle
messages received in the background.

The final UI will allow the user to enable/disable the receipt of
notifications. However, this PR concerns only the

new controller and the build management for the creation of the SW.

The new controller will work in accordance with two other controllers
for managing authentication and for managing the triggers necessary to
receive notifications.

## **Related issues**

No related issues.

Fixes:

## **Manual testing steps**

At the moment, the necessary environment variables have not been
inserted into the build process. To test the branch locally, feel free
to ask me for the env to insert into your local .metamaskrc file

## **Screenshots/Recordings**

In this video, I show the interaction between a test server for sending
notifications and the extension:

https://www.loom.com/share/702fc6a8aea74510acdb1bfac614fa4a

### **Before**

### **After**

Check the entire flow here
https://www.figma.com/file/c7GgNw2kScGrVyRGAPhwEd/Notifications?type=design&node-id=484%3A44713&mode=design&t=QsKrmdqZnx42IrMt-1

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've clearly explained what problem this PR is solving and how it
is solved.
- [ ] I've linked related issues
- [ ] I've included manual testing steps
- [x] I've included screenshots/recordings if applicable
- [ ] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [x] I’ve properly set the pull request status:
  - [x] In case it's not yet "ready for review", I've set it to "draft".
- [ ] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [x] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [x] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: Prithpal Sooriya <prithpal.sooriya@consensys.net>