chore(deps): bump keyring controller to ^15.0.0 (#24199)
## **Description**

This PR bumps the `metamask/keyring-controller` to `^15.0.0` and
addresses a bug in the keyring controller where invoking addNewAccount
inadvertently retrieves addresses from all keyrings, rather than
exclusively from the HD keyring. This behavior leads to inconsistencies
and errors when accounts from different keyring types are present. The
primary symptom of this bug is the retrieval of an incorrect number of
accounts when adding a new HD account. When the total number of accounts
in the client exceeds the number of HD accounts, it would cause the
retrieval of an account that does not exist yet on the HD keyring.

## **Related issues**

Fixes:
[23036](https://github.com/MetaMask/metamask-extension/issues/23036)
Fixes:
[21461](https://github.com/MetaMask/metamask-extension/issues/21461)

## **Manual testing steps**

1. Import 2 private key accounts
2. Add a new HD account and see that it is now successfully added. 

## **Screenshots/Recordings**

Not applicable

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: MetaMask Bot <metamaskbot@users.noreply.github.com>
Co-authored-by: legobeat <109787230+legobeat@users.noreply.github.com>
Co-authored-by: Gustavo Antunes <17601467+gantunesr@users.noreply.github.com>