feat: set useRequestQueue on by default (#24274)
This PR is to ensure that `useRequestQueue` is set to `true` by default.
After setting it to true, it turn the setting on for all users while
keeping the toggle, since this is a state update. This PR also includes
a migration.

## **Related issues**

Fixes: #23985 

## **Manual testing steps**

1. Go to settings > Experimental
2. Check `Select networks for each site` toggle is on
3. Toggle should work properly

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**


![Screenshot 2024-04-29 at 12 43
31 PM](https://github.com/MetaMask/metamask-extension/assets/39872794/6bdc6f49-6f20-44f1-a063-f9edd1dba09b)

### **After**

![Screenshot 2024-04-29 at 12 43
03 PM](https://github.com/MetaMask/metamask-extension/assets/39872794/19fce2fa-d0cb-43b8-8731-7b64c6d33975)



## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: Alex <adonesky@gmail.com>