feat: add support dynamic actions in alert modal (#24214)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
This PR aims to include an options property in `Alerts` called `actions`
which enables specific actions depending on the alert.
Updated `AlertModal` component to support displaying suitable buttons
based on the alert actions.
Created hooks to support the handling of these actions. 
<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

**Context**
Currently, the alert modal components display a single `Got it` button
that is used to close the modal or progress to the next alert in the
case of the `MultipleAlertModal`.

In addition to this button, some alerts can define additional actions
for the user to carry out, each of which will require an additional
button in the alert modal.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/24214?quickstart=1)

## **Related issues**

Fixes: https://github.com/MetaMask/MetaMask-planning/issues/2322

## **Manual testing steps**

Those components are currently not in use and are not available to
users. It is part of the new alert system, and we will have a dedicated
PR to integrate this component into the new confirmation screens.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**


![image](https://github.com/MetaMask/metamask-extension/assets/45455812/44c7433a-b59a-4400-bbf6-c6ab08e555f3)

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: Matthew Walsh <matthew.walsh@consensys.net>
Co-authored-by: George Marshall <george.marshall@consensys.net>