fix: Import all scripts fix (#24612)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

Fixes https://github.com/MetaMask/metamask-extension/issues/24514.

The problem was that if the service worker is terminated, when
app-init.js runs again when the service worker is to be restarted,
`importAllScripts()` is not called. The solution is just to make sure it
is called in the top-level scope, on every run of the script. This
should be safe to do because of the `scriptsLoadInitiated` check already
in the `importAllScripts` function.

The PR also uses the `serviceworker.state` property to ensure that we
don't attempt to import scripts before installation. (That is explained
further in the code comment added in this PR)

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/24612?quickstart=1)

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-extension/issues/24514

## **Manual testing steps**

1. Start MM MV3 build
2. Go to chrome://serviceworker-internals/?devtools
3. Stop Service Worker manually with the button
4. You should be able to open MetaMask

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.