fix: Add scuttle exception for fetch because it is used by chromium image util code when creating notifications (#24631)
## **Description**

MV3 bug that we needto fix:
https://github.com/MetaMask/metamask-extension/issues/24518

In particular, we've got:
```
LavaMoat - property "fetch" of globalThis is inaccessible under scuttling mode.
at get (chrome-extension://lbnhnppncgilfagdfbanifbfolodkdaf/scripts/runtime-lavamoat.js:13190:11)
  at loadImageDataForServiceWorker (extensions::imageUtil:24:22)
  at loadImageData (extensions::imageUtil:111:5)
  at Object.loadAllImages (extensions::imageUtil:134:5)
  at replaceNotificationOptionURLs (extensions::notifications:89:13)
  at extensions::notifications:116:5
  at chrome-extension://lbnhnppncgilfagdfbanifbfolodkdaf/common-6.js:3:431145
  at new Promise (<anonymous>)
  at Proxy.<anonymous> (chrome-extension://lbnhnppncgilfagdfbanifbfolodkdaf/common-6.js:3:430825)
  at Object.apply (chrome-extension://lbnhnppncgilfagdfbanifbfolodkdaf/common-6.js:3:430244)
  at Object._showNotification (chrome-extension://lbnhnppncgilfagdfbanifbfolodkdaf/common-2.js:1:455026)
```

The problematic code is in not in our codebase but in chromium, here:
https://github.com/chromium/chromium/blob/main/extensions/renderer/resources/image_util.js#L23-L25

You can see that function is named loadImageDataForServiceWorker and it
is only used in mv3.

This breaks the browser notification feature.

As can be seen from the commit history and the comments in this PR, we
attemtped some solutions that would allow us to provide a modified and
tamed fetch to be used for this image fetching. However, there are
unknown problems - related to how sentry uses fetch and how sentry is
initialized - that interfere with the solutions we attempted. So for
now, as discussed with @weizman, the latest commit on the PR just adds a
scuttling exception for fetch (and for Offscreen Canvas, which is also
needed for the browser notification feature).

We will return to the root problem later, and aim to remove these
scuttling exceptions.

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-extension/issues/24518

## **Manual testing steps**

1. `yarn dist:mv3`
2. install, onboard
3. send a transaction
4. verify that a browser notification is shown once the transaction is
confirmed

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.