fix: fix the counter in the header (#24733)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

This PR introduces a fix for the management of the counter in the menu.
Currently, the counter does not work for Snap notifications if wallet
notifications are not enabled.
The PR revises the logic we use to manage the count to be displayed by
using a hook (instead of selectors) and verifying in the header when to
show the counter.
In this way, the user has the ability to see the presence of received
Snap notifications regardless of whether they have enabled wallet
notifications.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/24733?quickstart=1)

## **Related issues**

Fixes:

## **Manual testing steps**

In the case of unread notifications, a small counter with a red
background is visible above the three dots for opening the menu,
indicating the number of unread notifications.​

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<img width="444" alt="Screenshot 2024-05-23 at 08 59 50"
src="https://github.com/MetaMask/metamask-extension/assets/1284304/19d5a57b-c5b1-491b-bcf8-3b730fd92fbc">

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [x] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [x] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.