fix: Place permission middleware ahead of all method implementations (#24472)
Adds all unrestricted RPC methods to the `unrestrictedMethods` array
passed to the permission controller, and moves the permission middleware
ahead of all RPC method implementations in `setupProviderEngine`. This
forces us to add all methods that aren't permission to the
`unrestrictedMethods` array, as a "method not found" error will be
thrown otherwise. It's safer this way.

Due the requirement that it swallows authorization errors, the
implementation of `eth_accounts` (a restricted method) is moved to
a new "legacy method middleware", which we by necessity place
ahead of the permission middleware. In addition, the responsibility of
rejecting intentionally unsupported RPC methods has been moved to
a dedicated middleware in front of the legacy method middleware.