chore: Bump `@metamask/permission-controller` to `9.1.0` (#24834)
Bumps `@metamask/permission-controller` to `9.1.0`, a minor version
bump.

This version introduces incremental permission requesting and granting,
as well as caveat mergers. A caveat merger function is implemented for
`restrictReturnedAccounts`. Meanwhile, the `eth_accounts` factory
function is modified such that it will use caveats specified via the
`permissionOptions` param if no `requestData` is present.