feat: Use sandboxed pages for Snaps execution in MV3 (#25171)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

Moves the execution of Snaps to a [sandboxed
page](https://developer.chrome.com/docs/extensions/how-to/security/sandboxing-eval)
referencing a local HTML file inside the offscreen document when using
the MV3 build. This effectively gives us access to `eval` without
hitting the network and should reduce the overhead when booting a Snap
(+ allow for Snaps to execute when the user is offline).

To support this, this PR introduces some new changes to the manifest as
well as the build process. For the build process we simply copy the same
iframe bundle currently used in the hosted version to
`/dist/chrome/snaps`. For the manifest, we add a reference to the
sandboxed page and tweak the CSP of the sandbox to be as restrictive as
possible (the default is not very strict). Then we can simply point the
existing offscreen executor to use the local iframe instead of the
remote one.

Closes https://github.com/MetaMask/metamask-extension/issues/25250

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/25171?quickstart=1)