docs: Manual scenario for 'Show test networks' toggle (#23091)
## **Description**
This PR documents the manual test scenario for the 'Show test networks'
toggle. Below is a summary of the validations included in this manual
scenario:

- Validate that the 'Show test networks' toggle is present in the
Advanced settings and is off by default.
- Validate that the toggle can be turned on and off.
- Validate that when the toggle is turned on, the Networks dropdown also
shows the 'Show test networks' as on, and the test networks are visible.
- Validate that when the toggle is turned off, the Networks dropdown
reflects the 'Show test networks' as off.

## **Related issues**

Fixes: #21988 

## **Manual testing steps**

[Refer
](https://github.com/MetaMask/metamask-extension/blob/doc/manual-scenario-show-test-network-toggle-21988/test/manual-scenarios/settings/advanced/show-test-networks.md)-->
test/scenarios/17. settings/advanced/show-test-networks.csv

## **Screenshots/Recordings**

https://recordit.co/ROidqMIdI7

---------

Co-authored-by: CW <chris.wilcox@consensys.net>