fix: `vault decryption` broken tests due to update on window handling (#26074)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

This PR fixes the broken test `vault-decryption`. The test was broken
when the introduction to the new window handling has been merged [in
this PR](https://github.com/MetaMask/metamask-extension/pull/25362).
The problem is that we are not using the socket communication when we
are not `IN_TEST` build, and in the `vault decryption` test
specifically, we are using a production build.

Since this is an exception, the fix is to disable the socket
communication and do not use the window handling functions from the
custom driver that rely on the socket communication for getting the
windows.


Error:

```
[driver] Called 'waitUntilXWindowHandles' with arguments [2]
[driver] Called 'switchToWindowWithTitle' with arguments ["MetaMask"]
Failure on testcase: 'undefined', for more information see the artifacts tab in CI

Error: No client connected to ServerMochaToBackground
```

Ci:
https://app.circleci.com/pipelines/github/MetaMask/metamask-extension/93155/workflows/ea0f878b-d13d-4765-b1da-41eb4224f273/jobs/3469123

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/26074?quickstart=1)

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-extension/issues/26075

## **Manual testing steps**

1. Check vault job in this PR
https://app.circleci.com/jobs/github/MetaMask/metamask-extension/3469757

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

![Screenshot from 2024-07-24
09-24-34](https://github.com/user-attachments/assets/318b24d4-2729-4a1b-af0a-60164b3717c0)


## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.