chore: Integrate SnapInsightsController (#26411)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

This PR integrates the `SnapInsightsController` and ports the existing
insights functionality to use this controller instead of the existing
hooks. This replaces our existing implementations with a
`useInsightSnaps` hook that returns all insights for a given
confirmation ID directly from the UI state.

The intention is that this matches the existing implementation as much
as possible, with one caveat that the timing for loading the insights
might have changed. Each insight is now added to state as soon as the
Snap responds.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/26411?quickstart=1)

## **Related issues**

Closes: https://github.com/MetaMask/snaps/issues/2568

## **Manual testing steps**

1. Install one or more transaction insights Snaps
2. Use the test-dapp to trigger transaction confirmations
3. See the transaction insights still works (including the modal)
4. Install one or more signature insights Snaps
5. Use the test-dapp to trigger signature confirmations
6. See that signature insights still works (including the modal)