fix(AssetPrice): change `stroke-linecap` and `stroke-linejoin` properties to valid JS syntax (#26488)
I noticed a react warning about the `stroke-linecap` and
`stroke-linejoin` properties not being valid (they are supposed to be
`strokeLinecap` and `strokeLinejoin`). I've updated them in this PR. It
doesn't change anything except for the warning, as React was passing the
hyphenated values to the DOM anyway (now it just won't complain).

So this PR just serves to get rid of the warning.