perf: add Sentry tracing to CircleCI runs (#26588)
This will allow CircleCI runs to report information to Sentry. It was very difficult to achieve, because there were specific E2E tests in `test/e2e/tests/metrics/errors.spec.js` that could never pass with this code added.

Before this PR, there was no way to pass runtime flags to the extension during tests. (Well okay, more precisely, you could change certain things through Fixtures, and you could change certain things through Ganache parameters, but both of these didn't load until late in the app load. Sentry is one of the first things to load, and it needed a way to get runtime flags very early in the app load.)

The way this works is a bit of a hack, but it's the only solution we could come up with that would load early enough:

1. A global beforeEach hook in the E2E tests backs up the manifest.json file
2. The `helpers.withFixtures()` function calls `setManifestFlags()`, which reads in the manifest file and parses it
3. We alter the manifest with CircleCI environment variables and custom flags, then write it back to manifest.json
4. The test runs, and the built extension can call `getManifestFlags()` to get the custom flags
5. A global afterEach hook restores the backup copy of the manifest, so that the next test gets the normal manifest

Co-authored-by: David Murdoch <187813+davidmurdoch@users.noreply.github.com>
Co-authored-by: Matthew Walsh <matthew.walsh@consensys.net>