perf: preload `_locales/en/messages.json` so we don't have to wait as long for our JS to dynamically load it (#26556)
We always load the English version of `messages.json`. This PR makes use
of link `rel=preload` so that the browser will get the file ready for
use immediately -- well before our JavaScript file that will eventually
need it asks for it. This reduces the amount of waiting due to Network
Waterfall (even if the network is the local file system).