fix(btc): fix jazzicons generations (#27662)
## **Description**

The jazzicons were all the same for mainnet/testnet accounts. It was
probably due to the fact that the namespace being used was `eip155` for
all addresses, but Bitcoin addresses have a different format.

Here's the technical details:
1. The current "icon factory" being in used is the ethereum one:
-
https://github.com/MetaMask/metamask-extension/blob/develop/ui/components/ui/jazzicon/jazzicon.component.tsx#L25
-
https://github.com/MetaMask/metamask-extension/blob/develop/ui/components/ui/jazzicon/jazzicon.component.tsx#L64
  - `namespace` always defaults to `eip155`
2. The default constructor used for the ethereum factory uses the
`jsNumberForAddress` which is ethereum-specific (or more like,
"hex-specific" here):
-
https://github.com/MetaMask/metamask-extension/blob/develop/ui/helpers/utils/icon-factory.ts#L40
-
https://github.com/MetaMask/metamask-extension/blob/develop/ui/helpers/utils/icon-factory.ts#L150-L154
- It slices the first 2 characters (probably to remove the `0x` prefix)
+ `parseInt(addr, 16)` will only work hex-strings, but Bitcoin is not
using this format
- the `parseInt` here will only consider the first valid hex-characters
of its input
  
To fix this, we check for the current address used for the jazzicon and
change the namespace based on this.

Ideally, we would want to use the `InternalAccount` object directly, but
that would require quite a lot of changes, so for now we keep this
simple.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/27662?quickstart=1)

## **Related issues**

N/A

## **Manual testing steps**

1. `yarn start:flask`
2. Settings > Experimental > "Enable Bitcoin support"
3. Create a Bitcoin mainnet account
4. Create a Bitcoin testnet account
5. Check that both jazzicons are different for those 2 Bitcoin accounts
6. Remove your Bitcoin accounts
7. Re-create them
8. Re-check that jazzicons are the same than step 5
9. "Hard"-restart your extension
10. Re-check that jazzicons are the same than step 5

## **Screenshots/Recordings**

### **Before**

![Screenshot 2024-10-07 at 16 17
35](https://github.com/user-attachments/assets/0a2e28e8-a81d-4468-9261-f1b0c8c3f02d)

### **After**

![Screenshot 2024-10-07 at 16 18
57](https://github.com/user-attachments/assets/81d378c9-6941-4067-9022-660e7dffb7b2)
## **Pre-merge author checklist**

- [ ] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.