fix: flaky test `Vault Decryptor Page is able to decrypt the vault pasting the text in the vault-decryptor webapp` (#27921)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
The problem is that we are clicking an element when it's moving, making
the click take no effect.
In the onboarding steps, there is a carousel, where you can move from
one page to the other. The issue is that the classical methods
findElement or clickElement, can take place without having the correct
page fully visible (while it's moving), making the click, take no
effect.

To avoid our clicks taking no effect, we need to add a new method to
wait until an element is not moving. That's the condition we need before
clicking that element.

Note. adding this by default to the clickElement can be an overkill as
the check will delay all instances of clickElement. Since this only
happens in the onboarding, I decided to **not** add it by default in the
clickEelemtn for this reason. Only in the onboarding flow, we need to
make sure we wait for elements not moving.


![Screenshot from 2024-10-17
08-48-13](https://github.com/user-attachments/assets/3b4df3b3-8d3b-49b7-8de7-34082410cfdd)


![image](https://github.com/user-attachments/assets/dbd1d928-3c32-42c4-8e2a-006f1345a54e)


[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/27921?quickstart=1)

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-extension/issues/27922

## **Manual testing steps**

1. Check ci run here
https://app.circleci.com/pipelines/github/MetaMask/metamask-extension/106117/workflows/e0bb3426-1c21-481b-8c0b-4a417f149172/jobs/3963202

## **Screenshots/Recordings**





https://github.com/user-attachments/assets/1d00a2b6-3b03-482f-a6c1-c14fb7e5c318


## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.