perf: add React.lazy to the Routes (#28172)
## **Description**

> **Note:** I would really appreciate quick reviews here mostly based on
"did I break anything." It's not really worth our time to nit-pick about
the style or the details because every line of this is going to be
changed by me again in the near future. Still to do in later PRs that
are going to completely re-write what I did here:
> - Convert routes.component.js to TypeScript
> - Convert routes.component.js to Hooks
> - Convert routes.component.js to react-router v6
> - Eliminate routes.container.js
>
> All this has to be done in sequential PRs based on top of this one, so
while I wait for reviews I can only work on a non-Routes topic.

This is the first PR to add `React.lazy` to the codebase. The meat of
the changes are in `ui/pages/routes/routes.component.js`.

Every other file changed is just a reaction to those changes. Much of
this is caused by the fact that `React.lazy` only works on default
exports ([see the documentation
here](https://react.dev/reference/react/lazy)), so I had to change a few
named exports to default exports.

I don't think this PR has much of an impact on loading times, but it
sets up further work that should have an impact.

There was one component (`ConfirmTransaction`) that needed some extra
attention to work with `React.lazy`. I had to remove the
`history.replace(mostRecentOverviewPage);`, which I'm pretty sure was an
unnecessary statement, and remove the Unit Test related to it.

**UPDATE:** I changed all of the Integration Tests to convert `getBy...`
to `await findBy...` to wait for the `React.lazy` component to load.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/28172?quickstart=1)

## **Related issues**

Progresses: MetaMask/MetaMask-planning#3302

<!--## **Manual testing steps**
## **Screenshots/Recordings**
## **Pre-merge author checklist**
## **Pre-merge reviewer checklist**-->