chore: adds Solana support for the account overview (#28411)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

We added support for the Solana account overview. Now when we select a
Solana address the user will be able to see its details in the home
view.
Also since the overview is the same for SOL and BTC, in order to not
repeat components, we've renamed as "non-evm" the existing BTC ones, and
reused them.

![Screenshot 2024-11-13 at 13 53
42](https://github.com/user-attachments/assets/649c1b1c-2da2-4f12-a18d-3549d8739c0e)

## **Related issues**

Fixes:

## **Manual testing steps**

As of right now, manually testing is a bit complex, it needs to run the
snap manually and the extension, since we 1st need to publish a new
release to npm with more up to date work. The snap version we have in
npm is outdated and won't support this flow. That said, if you want to
go ahead and run locally the steps are the following:

1. Clone the [ Solana Snap
monorepo](https://github.com/MetaMask/snap-solana-wallet) and run it
locally with `yarn` and then `yarn start`
2. In the extension, at this branch, apply the following changes and run
the extension as flask:
```
At builds.yml add the solana feature to the flask build:
 
     features:
       - build-flask
       - keyring-snaps
+      - solana

At shared/lib/accounts/solana-wallet-snap.ts point the snap ID to the snap localhost:

-export const SOLANA_WALLET_SNAP_ID: SnapId = SolanaWalletSnap.snapId as SnapId;
+//export const SOLANA_WALLET_SNAP_ID: SnapId = SolanaWalletSnap.snapId as SnapId;
+export const SOLANA_WALLET_SNAP_ID: SnapId = "local:http://localhost:8080/";
```
3. Manually install the snap via the snap dapp at http://localhost:3000
4. Enable the Solana account via Settings > Experimental > Enable Solana
account
5. Create a Solana account from the account-list menu and see the
account overview of it

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [ ] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: MetaMask Bot <metamaskbot@users.noreply.github.com>
Co-authored-by: Charly Chevalier <charly.chevalier@consensys.net>