fix: Support decimal point on advanced gas modal on mac (#28869)
## **Description**

The main bug this PR fixes only happens on mac, not linux. It is the
following: when a user pressed `.` on the advanced gas modal and a
number after it, a zero would appear instead the `.` and any numbers
before it. This is because in the `onChange` in `NumericInput`, when the
user presses `.`, `e.target.value` is `""`, and the code was defaulting
the value that got passed to the on change handler to `0`. To fix the
bug, this PR removes that default, and necessarily the `parseInt` that
wrapped it.

The PR also adds the `0` default to a few places where the newly
possible `""` value would otherwise break execution.

This PR also prevents the user from setting decimal custom nonce values.

Finally, it prevents `,` to be registered when `allowDecimals` is set on
`<FormField />`, most notably for the gas limit input.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/28869?quickstart=1)

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-extension/issues/28843

## **Manual testing steps**

1. Go to this page...
2.
3.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [ ] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.