test: remove `ganacheServer` from the `testSuite` and specs (#30536)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
In this PR, we remove the `ganacheServer` variable from the `testSuite`,
as we now have the `localNodes` array, which we already expose on the
`testSuite`. That array contains all the localnodes that we have started
for that specific test, no matter if Ganache or Anvil, or several.
So we can directly use the localNodes array instead of the specific
`ganacheServer` variable.
This variable is removed from withFixtures as well, as we handle the
ganacheServer in the array.

Additionally, I've set the local balance from the page object
`check_localNodeBalanceIsDisplayed` to '25 ETH', this way we can omit
passing the node in the method, as in most cases that is already the
default balance. Only in the case we have another balance (ie. when we
have deployed a smart contract, or use a different SRP) we can then pass
the localNodes.


[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/30536?quickstart=1)

## **Related issues**

Fixes:

## **Manual testing steps**

1. Go to this page...
2.
3.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [ ] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.