ci: migrate chrome e2e from circleci to gha (soft launch) (#30720)
## **Description**

This is a soft launch of the migration, as discussed in
https://consensys.slack.com/archives/C08FULM2VHC/p1741964592769319

This means that the CircleCI tests will still be active for now, and if
the GHA tests fail, it's just a warning but doesn't block merging.

Configs to remove later are marked with a `# soft launch` comment.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/30720?quickstart=1)

## **Related issues**

<!--## **Manual testing steps**
## **Screenshots/Recordings**
## **Pre-merge author checklist**
## **Pre-merge reviewer checklist**-->

---------

Co-authored-by: Howard Braham <howrad@gmail.com>
Co-authored-by: Norbert Elter <72046715+itsyoboieltr@users.noreply.github.com>